Fix broken specifications for user age and profile
User age function had a bad spec. The spec uses a hard coded date to test the method. That's probably bad and dumb, but it's fixed.

The user profile feature had a capybara ambiguity that I resolved. Though, I probably didn't use the recommended method to do it, it does work.

Signed-off-by: Richard Glover <richard@gloverdesigns.com>