Change “spirit animal” to “patronus”
Several articles of late have focused on the cultural appropriation of the notion of spirit animals. In the interest of being inclusive, and still being awesome, we at SWSDI wanted to update our code base to be more sensitive to the sacred concepts of native peoples. We think the idea of a Patronus (a’la Harry Potter) is a much more accurate way to think about what that piece of our application means to us.

Signed-off-by: Richard Glover <richard@gloverdesigns.com>