Fix the Camp Appliction Edit form
The Camp Application *edit* form shouldn’t have stuff regarding the payment on the bottom. NOR should it have links to an index of all the camp applications with destroy links. Fixed the controller for that index page, got rid of the views for it, and changed the form partial to eliminate the payment info.

Signed-off-by: Richard Glover <richard@gloverdesigns.com>